Migrate to shared form for workout
