+++ package.json
@@ -18,6 +18,7 @@
     "@chakra-ui/react": "^2.8.2",
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
+    "@hookform/devtools": "^4.3.1",
     "@hookform/resolvers": "^3.3.4",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.6.0",
@@ -36,6 +37,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.49.3",
+    "react-hook-form-persist": "^3.0.0",
     "react-select": "^5.8.0",
     "superjson": "^2.2.1",
     "zod": "^3.22.4"
